Move configuration to the `Plug` definition (#10)
Background:
To allow usage of metatags in multiple BEAM applications running on the
same VM (as is quite common in umbrella applications) we'd have to
rework how the default metatags is defined. Earlier the metatags would
be configured globally which meant the configuration would
clash/overwrite each other. This change will move the defining of
metatags to be where the plug is used instead.

Configuring defaults will still be available through
`Application.get_env/3`.

Changes:
- Update README
- Allow the plug to take the default metadata